Add suggestions from review
