+++ CliWrap.Tests/ValidationSpecs.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using CliWrap.Buffered;
 using CliWrap.Exceptions;
+using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -55,6 +56,24 @@ namespace CliWrap.Tests
 
             // Act & assert
             await cmd.ExecuteAsync();
+        }
+
+        [Fact(Timeout = 15000)]
+        public async Task Command_with_non_zero_exit_code_yields_exit_code_in_exception()
+        {
+            // Arrange
+            var cmd = Cli.Wrap("dotnet")
+                .WithArguments(a => a
+                    .Add(Dummy.Program.FilePath)
+                    .Add("exit-with")
+                    .Add("--code").Add(1));
+
+            // Act
+            var task = cmd.ExecuteAsync();
+
+            // Assert
+            var ex = await Assert.ThrowsAsync<CommandExecutionException>(async () => await task);
+            ex.ExitCode.Should().Be(1);
         }
     }
 }
